// src/system/types.ts
export type QualityLabel = 'sd' | 'hd' | '1080' | '2k' | '4k' | '8k';

export interface Download {
    type: 'AudioVideo' | 'AudioOnly' | 'VideoOnly';
    id: string; // Unique identifier for the download
    title?: string;
    quality: string;
    url: string;
    format: string;
    filesize?: number;
    videoQuality?: { label: QualityLabel; fps?: number };
    width?: number;
    height?: number;
}

export interface Thumbnail {
    label: string;
    url: string;
    width?: number;
    height?: number;
}

interface Subtitle {
    languageCode: string;
    languageName?: string;
    url: string;
    format: string;
    autoGenerated?: boolean;
}

export interface SingleInfo {
    type: 'single';
    id: string;
    title: string;
    length: number;
    description?: string;
    author?: string;
    downloads: Download[];
    thumbnailUrl?: string;
    thumbnails?: Thumbnail[];
    subtitles?: Subtitle[];
}
// ...
export type Mp3Bitrate = '128' | '192' | '256' | '320';